Extract importer file index module
