+++ packages/reprint-importer/src/lib/index/class-index-line-path-key-extractor.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Reprint\Importer\Index;
+
+final class IndexLinePathKeyExtractor
+{
+    public function __invoke(string $line): ?string
+    {
+        $entry = IndexLineParser::parse($line);
+        return $entry !== null ? $entry["path"] : null;
+    }
+}
