+++ packages/reprint-importer/src/lib/index/load.php
@@ -1,0 +1,7 @@
+<?php
+
+require_once __DIR__ . '/class-index-line-parser.php';
+require_once __DIR__ . '/class-index-line-path-key-extractor.php';
+require_once __DIR__ . '/class-index-path-prefix-matcher.php';
+require_once __DIR__ . '/class-index-file-sorter.php';
+require_once __DIR__ . '/class-index-store.php';
